refactor: 날짜 구하는 로직 utils로 파일 분리
