+++ components/community/Board/New/index.tsx
@@ -7,24 +7,13 @@ import { postContentState, postListState, postTitleState } from "recoil/states";
 import Editor from "./Editor/index";
 import HeadingText from "../../../common/HeadingText";
 
+import { getDetailDate } from "utils/date";
+
 import * as S from "./style";
 
 let id = 4;
 const getId = () => {
   return id++;
-};
-
-const getDate = () => {
-  const date = new Date();
-  const year = String(date.getFullYear());
-  const month = String(date.getMonth() + 1).padStart(2, "0");
-  const day = String(date.getDate()).padStart(2, "0");
-  const hour = String(date.getHours()).padStart(2, "0");
-  const minute = String(date.getMinutes()).padStart(2, "0");
-  const second = String(date.getSeconds()).padStart(2, "0");
-  const mSecond = String(date.getMilliseconds()).padStart(3, "0");
-
-  return year + month + day + hour + minute + second + mSecond;
 };
 
 const New = () => {
@@ -33,7 +22,6 @@ const New = () => {
   const [title, setTitle] = useRecoilState(postTitleState);
   const [content, setContent] = useRecoilState(postContentState);
   const setPostList = useSetRecoilState(postListState);
-  console.log(getDate());
 
   /* 게시글 작성 페이지 - 저장 버튼 클릭 처리 함수 */
   const handleSubmit = () => {
@@ -45,7 +33,7 @@ const New = () => {
       {
         id: getId(),
         writer: "임하림",
-        uploadTime: getDate(),
+        uploadTime: getDetailDate(),
         title: title,
         content: content,
         views: 30,
